Remove RecoLocalCalo/EcalDeadChannelRecoveryProducers and RecoCaloTools/EcalChannelKiller since they are only for testing purposes, move the contents to RecoLocalCalo/EcalDeadChannelRecoveryAlgos/test/
